refactor: extract shared batch execution loop into lib/batch-exec
Eliminates ~110 lines of duplicated timeout-budget + sequential-exec
logic between the HTTP route and MCP tool by consolidating into a
single executeBatch() helper.
